+++ src/layouts/TokenDetail.svelte
